Feature/unit testing to dev

+++ src/main/java/com/qa/tdl/rest/AssigneeController.java
@@ -16,8 +16,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.qa.tdl.persistance.domain.AssigneeDomain;
-import com.qa.tdl.persistance.dtos.AssigneeDTO;
+import com.qa.tdl.persistence.domain.AssigneeDomain;
+import com.qa.tdl.persistence.dtos.AssigneeDTO;
 import com.qa.tdl.services.AssigneeService;
 
 @RestController
@@ -39,7 +39,7 @@ public class AssigneeController {
 	}
 	
 	@GetMapping("/read/{id}")
-	public ResponseEntity<AssigneeDTO> readAssignee(@PathVariable int id) {
+	public ResponseEntity<AssigneeDTO> readAssignee(@PathVariable long id) {
 		return ResponseEntity.ok(this.service.readAssignee(id));
 	}
 	
@@ -51,7 +51,7 @@ public class AssigneeController {
 	
 	// PUT
 	@PutMapping("/update")
-	public ResponseEntity<AssigneeDTO> updateAssignee(@PathParam("id") int id, @RequestBody AssigneeDomain model) {
+	public ResponseEntity<AssigneeDTO> updateAssignee(@PathParam("id") long id, @RequestBody AssigneeDomain model) {
 		return new ResponseEntity<>(this.service.update(id, model), HttpStatus.ACCEPTED);
 	}
